+++ iconify_full/lib/src/fast_cached/iconify_fast_cache_storage_web.dart
@@ -1,0 +1,17 @@
+Future<String?> readSvg(String? root, String prefix, String name) async =>
+    null;
+
+Future<void> writeSvg(
+  String? root,
+  String prefix,
+  String name,
+  String svg,
+) async {}
+
+Future<String?> readJson(String? root, String prefix) async => null;
+
+Future<void> writeJson(String? root, String prefix, String body) async {}
+
+Future<String?> readCollections(String? root) async => null;
+
+Future<void> writeCollections(String? root, String body) async {}